requirements.txt
aiohttp==3.7.4.post0
async-timeout==3.0.1
attrs==21.2.0
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
Deprecated==1.2.12
gitdb==4.0.7
github.py==0.5.0
GitPython==3.1.17
idna==2.10
idna-ssl==1.1.0
multidict==5.1.0
pycparser==2.20
PyGithub==1.55
PyJWT==2.1.0
PyNaCl==1.4.0
requests==2.25.1
six==1.16.0
smmap==4.0.0
typing-extensions==3.10.0.0
urllib3==1.26.5
wrapt==1.12.1
yarl==1.6.3
